requirements.txt
anthropic>=0.30.0
flask>=3.0.0
click>=8.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-docx>=1.1.0
